handling an exception

+++ src/main/java/com/philimonov/spring/rest/controller/MyRestController.java
@@ -51,6 +51,10 @@ public class MyRestController {
 
     @DeleteMapping("/employees/{id}")
     public String deleteEmployee(@PathVariable int id) {
+        Employee employee = employeeService.getEmployee(id);
+        if(employee == null){
+            throw new NoSuchEmployeeException("There is no employee with ID = " + id + " in Database.");
+        }
         employeeService.deleteEmployee(id);
         return "Employee with id = " + id + " was deleted!";
     }
